Bug/oracle auto limit plus refactor (#1)
* refactoring

* add limit text for oracle

* simplify logic

* add tests for oracle

* move tests to query runner

* remove unused tests
