+++ redash/query_runner/oracle.py
@@ -33,6 +33,8 @@ logger = logging.getLogger(__name__)
 class Oracle(BaseSQLQueryRunner):
     should_annotate_query = False
     noop_query = "SELECT 1 FROM dual"
+    limit_query = " FETCH NEXT 1000 ROWS ONLY"
+    limit_keywords = [ "ROW", "ROWS", "ONLY", "TIES"]
 
     @classmethod
     def get_col_type(cls, col_type, scale):
