+++ tests/query_runner/test_basesql_queryrunner.py
@@ -9,6 +9,38 @@ class TestBaseSQLQueryRunner(unittest.TestCase):
     def setUp(self):
         self.query_runner = BaseSQLQueryRunner({})
 
+    def test_check_query_limit_no_limit(self):
+        query = "SELECT *"
+        self.assertEqual(True, self.query_runner.query_is_select_no_limit(query))
+
+    def test_check_query_limit_non_select(self):
+        query = "Create Table (PersonID INT)"
+        self.assertEqual(False, self.query_runner.query_is_select_no_limit(query))
+
+    def test_check_query_limit_invalid_1(self):
+        query = "OFFSET 5"
+        self.assertEqual(False, self.query_runner.query_is_select_no_limit(query))
+
+    def test_check_query_limit_invalid_2(self):
+        query = "TABLE A FROM TABLE B"
+        self.assertEqual(False, self.query_runner.query_is_select_no_limit(query))
+
+    def test_check_query_with_limit(self):
+        query = "SELECT * LIMIT 5"
+        self.assertEqual(False, self.query_runner.query_is_select_no_limit(query))
+
+    def test_check_query_with_offset(self):
+        query = "SELECT * LIMIT 5 OFFSET 3"
+        self.assertEqual(False, self.query_runner.query_is_select_no_limit(query))
+    
+    def test_add_limit_query_no_limit(self):
+        query = "SELECT *"
+        self.assertEqual("SELECT * LIMIT 1000", self.query_runner.add_limit_to_query(query))
+
+    def test_add_limit_query_with_punc(self):
+        query = "SELECT *;"
+        self.assertEqual("SELECT * LIMIT 1000;", self.query_runner.add_limit_to_query(query))
+
     def test_apply_auto_limit_origin_no_limit_1(self):
         origin_query_text = "SELECT 2"
         query_text = self.query_runner.apply_auto_limit(origin_query_text, True)
